fix(06-05): resolve incomplete type errors with std::optional
- Include mcpp/core/error.h in json_rpc.h for JsonRpcError complete type
- Reorder capability structs in capabilities.h to define before use in std::optional
- Cast size_t to int64_t in request_tracker.cpp for RequestId variant compatibility
- Make json-schema include optional with __has_include check

These are pre-existing bugs that prevented compilation. std::optional requires
complete types at instantiation point, not just forward declarations.

Rule 1 (Bug) auto-fixes for blocking compilation issues.

+++ src/mcpp/core/request_tracker.cpp
@@ -5,7 +5,7 @@ namespace mcpp::core {
 RequestId RequestTracker::next_id() {
     // Atomic increment with sequential consistency
     // Returns the value before increment (0, 1, 2, ...)
-    return counter_.fetch_add(1, std::memory_order_seq_cst);
+    return static_cast<int64_t>(counter_.fetch_add(1, std::memory_order_seq_cst));
 }
 
 void RequestTracker::register_pending(
